Update the terraform gitignore file
* The tfstate backup file can sometimes be named foo.tfstate.${timestamp or
hash}.backup and there is also a tfstate.d/ directory that should also
be ignored. *.tfstate.* catches those two scenarios

* I've removed the patch file and merged that line back in to the main
template file since it's something that should be (encouraged to be)
ignored by default when working with terraform. I know the patch file
added the line back in but as stated in the blog https://blog.joeblau.com/gitignore-io-template-fork
the intention is to get use patch files less. In this case the terraform
template is not really neglected since the tool is fairly new(ish).